Try to have a better crud
